+++ src/stores/sized.rs
@@ -1,10 +1,10 @@
+use super::Cached;
+use crate::lru_list::LRUList;
 use hashbrown::raw::RawTable;
 use std::cmp::Eq;
 use std::collections::hash_map::RandomState;
 use std::fmt;
 use std::hash::{BuildHasher, Hash, Hasher};
-
-use super::{Cached, LRUList};
 
 #[cfg(feature = "async")]
 use {super::CachedAsync, async_trait::async_trait, futures::Future};
