+++ src/stores/timed.rs
@@ -1,10 +1,9 @@
+use super::Cached;
 use std::cmp::Eq;
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::hash::Hash;
 use std::time::Instant;
-
-use super::Cached;
 
 #[cfg(feature = "async")]
 use {super::CachedAsync, async_trait::async_trait, futures::Future};
